dcp-815 rename ampq file to stop shadowing kombu
